ui_tests/requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
h11==0.12.0
idna==3.3
lxml==4.6.5
outcome==1.1.0
paramiko==2.8.0
pycparser==2.20
PyNaCl==1.4.0
pyOpenSSL==21.0.0
requests==2.26.0
robotframework==4.1.2
robotframework-pythonlibcore==3.0.0
robotframework-seleniumlibrary==5.1.3
robotframework-sshlibrary==3.7.0
scp==0.14.1
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
webdrivermanager==0.10.0
wsproto==1.0.0
